+++ src/CosmoFuse/visualisation.py
@@ -1,7 +1,6 @@
 import matplotlib.pyplot as plt
 import numpy as np
 from scipy.ndimage import gaussian_filter
-from scipy.stats import gaussian_kde
 
 
 def contours(